EVO-21 - Rename some directories in OrderDetailPage for readability.
