--- app/containers/OrderPage/OrderDetailPage/OrderDetail/index.tsx
@@ -1,33 +1,0 @@
-/*
- * orderDetails
- *
- * This is a the details of an order
- *
- */
-
-// remove, use babel for antD css
-import { Col, Layout, Row } from 'antd';
-import 'antd/dist/antd.css';
-
-import * as React from 'react';
-import OrderDetailHeader from '../OrderDetailHeader';
-import PracticeInformation from '../OrderDetailPractice';
-import OrderDetailClient from '../OrderDetailClient';
-
-
-export default function OrderDetail() {
-
-  return (
-    <Layout>
-      <OrderDetailHeader/>
-      <Row gutter={16} align="top" style={{ marginTop: '20px' }}>
-        <Col className="gutter-row" span={6}>
-          <PracticeInformation/>
-        </Col>
-        <Col className="gutter-row" span={6}>
-          <OrderDetailClient/>
-        </Col>
-      </Row>
-    </Layout>
-  );
-}
