+++ app/containers/OrderPage/OrderDetailPage/PracticeInformation/constants.tsx
