+++ app/containers/OrderPage/OrderDetailPage/PracticeInformation/index.tsx
