+++ app/containers/OrderPage/OrderDetailPage/ShippingInformation/constants.tsx
@@ -1,0 +1,4 @@
+export const shippingDetailData = [
+  ['Method:', 'Next Day'],
+  ['Address:', '14233 N 50th Ln Glendale, AZ 85306'],
+];
